key_grabber: call XResizeWindow when pulling down
In a setting with XMonad and multiple monitors with different
resolutions, when pulling down from a smaller monitor, the tilda window
size becomes smaller and smaller. This adds a resize function after the
call to fix the bug (might be gtk bug, though).

Note that XMonad is not tiling tilda but merely managing it.

Reproduction of the bug:
* Use XMonad as the window manager, set Tilda to floating
* Press the pulldown button the smaller screen which doesn't have tilda
* Expected outcome: tilda pulls down with its original size
* Actual outcome: tilda pulls down with the size proportionally smaller
  according to the smaller screen.